requirements_api.txt
flask==3.0.0
flask-cors==4.0.0
redis==5.0.1
numpy==1.24.4
implicit==0.7.0
python-dotenv==1.0.0
scipy>=1.10.0,<1.12.0
pandas>=1.5.0,<2.0.0
psycopg2-binary>=2.9.0


